+++ ui/app/src/formik/DNSChaos.ts
@@ -17,6 +17,12 @@ export const actions = ['error', 'random'],
       label: 'patterns',
       value: [],
       helperText:
-        'Optional. Choose which domain names to take effect, support the placeholder ? and wildcard *, or the Specified domain name. Note:      1. The wildcard * must be at the end of the string. For example, chaos-*.org is invalid.      2. if the patterns is empty, will take effect on all the domain names. For example:   The value is ["google.com", "github.*", "chaos-mes?.org"],   will take effect on "google.com", "github.com" and "chaos-mesh.org"',
+        'Optional. Choose which domain names to take effect, support the placeholder ? and wildcard *, or the Specified domain name. Note: 1. The wildcard * must be at the end of the string. For example, chaos-*.org is invalid. 2. if the patterns is empty, will take effect on all the domain names. For example: The value is ["google.com", "github.*", "chaos-mes?.org"], will take effect on "google.com", "github.com" and "chaos-mesh.org"',
+    },
+    {
+      field: 'text',
+      label: 'remoteCluster',
+      value: '',
+      helperText: 'Optional. RemoteCluster represents the remote cluster where the chaos will be deployed',
     },
   ]