+++ ui/app/src/formik/GCPChaos.ts
@@ -26,6 +26,12 @@ export const actions = ['node-stop', 'node-reset', 'disk-loss'],
     },
     {
       field: 'text',
+      label: 'remoteCluster',
+      value: '',
+      helperText: 'Optional. RemoteCluster represents the remote cluster where the chaos will be deployed',
+    },
+    {
+      field: 'text',
       label: 'secretName',
       value: '',
       helperText: 'Optional. SecretName defines the name of kubernetes secret. It is used for GCP credentials.',