+++ ui/app/src/pages/Dashboard/Predefined.tsx
@@ -15,13 +15,12 @@
  *
  */
 import loadable from '@loadable/component'
-
 import { Box, Button, Card, Modal, Typography } from '@mui/material'
 import { makeStyles } from '@mui/styles'
 import { Ace } from 'ace-builds'
-import api from 'api'
 import clsx from 'clsx'
 import yaml from 'js-yaml'
+import { postExperiments, postSchedules } from 'openapi'
 import { useEffect, useRef, useState } from 'react'
 import { useIntl } from 'react-intl'
 
@@ -112,9 +111,7 @@ const Predefined = () => {
     const exp: any = yaml.load(yamlEditor!.getValue())
 
     const isSchedule = exp['kind'] === 'Schedule'
-    const action = isSchedule
-      ? (schedule: any) => api.schedules.schedulesPost({ schedule })
-      : (chaos: any) => api.experiments.experimentsPost({ chaos })
+    const action = isSchedule ? (schedule: any) => postSchedules(schedule) : (chaos: any) => postExperiments(chaos)
 
     action(exp)
       .then(() => {